+++ src/app/(app)/@modals/params.tsx
@@ -2,6 +2,7 @@ export enum ModalParam {
   AddQuest = "add-quest",
   BuildInfo = "build-info",
   QuestChains = "quest-chains",
+  QuestTypes = "quest-types",
   Settings = "settings",
   Stats = "stats",
   User = "user",